[refactor] 필요 없는 코드 제거, 네이밍 수정
