--- Routinus/Routinus/Presentation/Home/Cells/CalendarHeader.swift
@@ -1,209 +1,0 @@
-//
-//  CalendarHeader.swift
-//  Routinus
-//
-//  Created by 김민서 on 2021/11/02.
-//
-
-import UIKit
-
-final class CalendarHeader: UIView {
-    lazy var titleLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = .systemFont(ofSize: 18, weight: .semibold)
-        label.text = "요약"
-        label.accessibilityTraits = .header
-        label.isAccessibilityElement = true
-        return label
-    }()
-
-    lazy var monthLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.font = .systemFont(ofSize: 24, weight: .bold)
-        label.text = "Month"
-        label.accessibilityTraits = .header
-        label.isAccessibilityElement = true
-        return label
-    }()
-
-    lazy var previousMonthButton: UIButton = {
-        let button = UIButton()
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.titleLabel?.font = .systemFont(ofSize: 17, weight: .medium)
-        button.titleLabel?.textAlignment = .left
-
-        if let chevronImage = UIImage(systemName: "chevron.left") {
-            let imageAttachment = NSTextAttachment(image: chevronImage)
-            let attributedString = NSMutableAttributedString()
-
-            attributedString.append(NSAttributedString(attachment: imageAttachment))
-            attributedString.append(NSAttributedString(string: ""))
-
-            button.setAttributedTitle(attributedString, for: .normal)
-        } else {
-            button.setTitle("Previous", for: .normal)
-        }
-
-        button.titleLabel?.textColor = .label
-
-        button.addTarget(self, action: #selector(didTapPreviousMonthButton), for: .touchUpInside)
-        return button
-    }()
-
-    lazy var nextMonthButton: UIButton = {
-        let button = UIButton()
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.titleLabel?.font = .systemFont(ofSize: 17, weight: .medium)
-        button.titleLabel?.textAlignment = .right
-
-        if let chevronImage = UIImage(systemName: "chevron.right") {
-            let imageAttachment = NSTextAttachment(image: chevronImage)
-            let attributedString = NSMutableAttributedString(string: "")
-
-            attributedString.append(NSAttributedString(attachment: imageAttachment))
-
-            button.setAttributedTitle(attributedString, for: .normal)
-        } else {
-            button.setTitle("Next", for: .normal)
-        }
-
-        button.titleLabel?.textColor = .label
-
-        button.addTarget(self, action: #selector(didTapNextMonthButton), for: .touchUpInside)
-        return button
-    }()
-
-    lazy var dayOfWeekStackView: UIStackView = {
-        let stackView = UIStackView()
-        stackView.translatesAutoresizingMaskIntoConstraints = false
-        stackView.distribution = .fillEqually
-        return stackView
-    }()
-
-    lazy var separatorView: UIView = {
-        let view = UIView()
-        view.translatesAutoresizingMaskIntoConstraints = false
-        view.backgroundColor = UIColor.label.withAlphaComponent(0.2)
-        return view
-    }()
-
-    private lazy var dateFormatter: DateFormatter = {
-        let dateFormatter = DateFormatter()
-        dateFormatter.calendar = Calendar(identifier: .gregorian)
-        dateFormatter.locale = Locale.autoupdatingCurrent
-        dateFormatter.setLocalizedDateFormatFromTemplate("y.MMM")
-        return dateFormatter
-    }()
-
-    var baseDate = Date() {
-        didSet {
-            monthLabel.text = dateFormatter.string(from: baseDate)
-        }
-    }
-
-    let didTapLastMonthCompletionHandler: (() -> Void)
-    let didTapNextMonthCompletionHandler: (() -> Void)
-
-    init(
-        didTapLastMonthCompletionHandler: @escaping (() -> Void),
-        didTapNextMonthCompletionHandler: @escaping (() -> Void)
-    ) {
-        self.didTapLastMonthCompletionHandler = didTapLastMonthCompletionHandler
-        self.didTapNextMonthCompletionHandler = didTapNextMonthCompletionHandler
-
-        super.init(frame: CGRect.zero)
-
-        translatesAutoresizingMaskIntoConstraints = false
-
-        backgroundColor = .systemGroupedBackground
-
-        layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
-        layer.cornerCurve = .continuous
-        layer.cornerRadius = 15
-
-        addSubview(titleLabel)
-        addSubview(monthLabel)
-        addSubview(previousMonthButton)
-        addSubview(nextMonthButton)
-        addSubview(dayOfWeekStackView)
-        addSubview(separatorView)
-
-        for dayNumber in 1...7 {
-            let dayLabel = UILabel()
-            dayLabel.font = .systemFont(ofSize: 12, weight: .bold)
-            dayLabel.textColor = .secondaryLabel
-            dayLabel.textAlignment = .center
-            dayLabel.text = dayOfWeekLetter(for: dayNumber)
-
-            dayLabel.isAccessibilityElement = false
-            dayOfWeekStackView.addArrangedSubview(dayLabel)
-        }
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    private func dayOfWeekLetter(for dayNumber: Int) -> String {
-        switch dayNumber {
-        case 1:
-            return "S"
-        case 2:
-            return "M"
-        case 3:
-            return "T"
-        case 4:
-            return "W"
-        case 5:
-            return "T"
-        case 6:
-            return "F"
-        case 7:
-            return "S"
-        default:
-            return ""
-        }
-    }
-
-    override func layoutSubviews() {
-        super.layoutSubviews()
-
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: topAnchor, constant: 15),
-            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 20),
-
-            nextMonthButton.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -10),
-            nextMonthButton.topAnchor.constraint(equalTo: topAnchor, constant: 15),
-
-            monthLabel.topAnchor.constraint(equalTo: topAnchor, constant: 15),
-            monthLabel.trailingAnchor.constraint(equalTo: nextMonthButton.leadingAnchor, constant: -10),
-
-            previousMonthButton.trailingAnchor.constraint(equalTo: monthLabel.leadingAnchor, constant: -10),
-            previousMonthButton.topAnchor.constraint(equalTo: topAnchor, constant: 15),
-
-            dayOfWeekStackView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            dayOfWeekStackView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            dayOfWeekStackView.bottomAnchor.constraint(equalTo: separatorView.bottomAnchor, constant: -5),
-
-            separatorView.leadingAnchor.constraint(equalTo: leadingAnchor),
-            separatorView.trailingAnchor.constraint(equalTo: trailingAnchor),
-            separatorView.bottomAnchor.constraint(equalTo: bottomAnchor),
-            separatorView.heightAnchor.constraint(equalToConstant: 1)
-        ])
-
-        let smallDevice = UIScreen.main.bounds.width <= 350
-        let fontPointSize: CGFloat = smallDevice ? 14 : 17
-        previousMonthButton.titleLabel?.font = .systemFont(ofSize: fontPointSize, weight: .medium)
-        nextMonthButton.titleLabel?.font = .systemFont(ofSize: fontPointSize, weight: .medium)
-    }
-
-    @objc func didTapPreviousMonthButton() {
-        didTapLastMonthCompletionHandler()
-    }
-
-    @objc func didTapNextMonthButton() {
-        didTapNextMonthCompletionHandler()
-    }
-}
