+++ Routinus/Routinus/Presentation/Home/Subviews/CalendarView.swift
@@ -25,18 +25,18 @@ final class CalendarView: UIView {
     }()
 
     private lazy var headerView = CalendarHeader(
-        didTapLastMonthCompletionHandler: { [weak self] in
+        didTappedLastMonthCompletionHandler: { [weak self] in
             guard let self = self else { return }
             self.viewModel?.changeDate(month: -1)
             self.calendarView.reloadData()
         },
-        didTapNextMonthCompletionHandler: { [weak self] in
+        didTappedNextMonthCompletionHandler: { [weak self] in
             guard let self = self else { return }
             self.viewModel?.changeDate(month: 1)
             self.calendarView.reloadData()
         })
 
-    private var viewModel: HomeViewModelIO? = nil
+    private var viewModel: HomeViewModelIO?
     private var cancellables = Set<AnyCancellable>()
 
     weak var delegate: UICollectionViewDelegateFlowLayout? {
@@ -53,40 +53,20 @@ final class CalendarView: UIView {
 
     init(viewModel: HomeViewModelIO?) {
         self.viewModel = viewModel
-        super.init(frame: CGRect(origin: CGPoint(x: 0, y: 0), size: CGSize(width: 0, height: 0)))
+        super.init(frame: CGRect(origin: .zero, size: .zero))
         self.configureView()
         self.configureViewModel()
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        self.configureView()
     }
 
     required init?(coder: NSCoder) {
         super.init(coder: coder)
         self.configureView()
+        self.configureViewModel()
     }
 
     func configureView() {
         self.addSubview(calendarView)
         self.addSubview(headerView)
-
-        var constraints = [
-            headerView.leadingAnchor.constraint(equalTo: self.readableContentGuide.leadingAnchor),
-            headerView.trailingAnchor.constraint(equalTo: self.readableContentGuide.trailingAnchor),
-            headerView.heightAnchor.constraint(equalToConstant: 85),
-            headerView.topAnchor.constraint(equalTo: self.readableContentGuide.topAnchor)
-        ]
-
-        constraints.append(contentsOf: [
-            calendarView.leadingAnchor.constraint(equalTo: self.headerView.leadingAnchor),
-            calendarView.trailingAnchor.constraint(equalTo: self.headerView.trailingAnchor),
-            calendarView.topAnchor.constraint(equalTo: self.headerView.bottomAnchor),
-            calendarView.heightAnchor.constraint(equalTo: self.heightAnchor, multiplier: 0.8)
-        ])
-
-        NSLayoutConstraint.activate(constraints)
 
         calendarView.register(
             CalendarDateCell.self,
@@ -96,8 +76,20 @@ final class CalendarView: UIView {
         headerView.baseDate = self.viewModel?.baseDate.value ?? Date()
     }
 
-    func reloadData() {
-        self.calendarView.reloadData()
+    override func layoutSubviews() {
+        super.layoutSubviews()
+
+        NSLayoutConstraint.activate([
+            headerView.leadingAnchor.constraint(equalTo: self.readableContentGuide.leadingAnchor),
+            headerView.trailingAnchor.constraint(equalTo: self.readableContentGuide.trailingAnchor),
+            headerView.heightAnchor.constraint(equalToConstant: 85),
+            headerView.topAnchor.constraint(equalTo: self.readableContentGuide.topAnchor),
+
+            calendarView.leadingAnchor.constraint(equalTo: self.headerView.leadingAnchor),
+            calendarView.trailingAnchor.constraint(equalTo: self.headerView.trailingAnchor),
+            calendarView.topAnchor.constraint(equalTo: self.headerView.bottomAnchor),
+            calendarView.heightAnchor.constraint(equalTo: self.heightAnchor, multiplier: 0.8)
+        ])
     }
 
     private func configureViewModel() {
@@ -109,5 +101,9 @@ final class CalendarView: UIView {
                 self.calendarView.reloadInputViews()
             })
             .store(in: &cancellables)
+    }
+
+    func reloadData() {
+        self.calendarView.reloadData()
     }
 }
